build: prepare portable Windows releases and upstream compatibility checks
Add CMake presets, static MSVC/SDL packaging with DLL auditing, build revision metadata, and Linux/Windows CI. Centralize production-state reads and document release and upstream maintenance workflows.

Validate Linux GCC/Clang tests, GUI smoke checks, compatibility-tool isolation, and firmware builds. Windows/MSVC and Windows 11 runtime verification remain pending.

+++ src/ui/DisplayFormatting.h
@@ -18,12 +18,14 @@ inline void formatTrip(char* text, size_t size, int64_t distanceMm) {
   if (meters < 10000ULL) {
     std::snprintf(text, size, "%s%" PRIu64 ".%03" PRIu64,
                   negative ? "-" : "",
-                  meters / 1000ULL, meters % 1000ULL);
+                  static_cast<uint64_t>(meters / 1000ULL),
+                  static_cast<uint64_t>(meters % 1000ULL));
   } else {
     const uint64_t tenMeters = meters / 10ULL;
     std::snprintf(text, size, "%s%" PRIu64 ".%02" PRIu64,
                   negative ? "-" : "",
-                  tenMeters / 100ULL, tenMeters % 100ULL);
+                  static_cast<uint64_t>(tenMeters / 100ULL),
+                  static_cast<uint64_t>(tenMeters % 100ULL));
   }
 }
